chore(base): Move `Room::*knock*` methods into the new `knock` module.
This patch moves the `Room::*knock*` methods into the new `knock`
module.

The idea is to group API by “theme” to get smaller modules and more
organised code.
